+++ source/mact/src/control.cpp
@@ -429,10 +429,10 @@ void CONTROL_ClearAssignments(void)
     memset(CONTROL_MouseAxesMap,        AXISUNDEFINED,   sizeof(CONTROL_MouseAxesMap));
     memset(CONTROL_MouseButtonMapping,  BUTTONUNDEFINED, sizeof(CONTROL_MouseButtonMapping));
 
-    for (int & i : CONTROL_MouseAxesScale)
+    for (int32_t & i : CONTROL_MouseAxesScale)
         i = NORMALAXISSCALE;
 
-    for (int & i : CONTROL_JoyAxesScale)
+    for (int32_t & i : CONTROL_JoyAxesScale)
         i = NORMALAXISSCALE;
 }
 
@@ -811,8 +811,8 @@ bool CONTROL_Startup(controltype which, int32_t(*TimeFunction)(void), int32_t ti
     CONTROL_MousePresent    = Mouse_Init();
     CONTROL_MouseEnabled    = CONTROL_MousePresent;
 
-    CONTROL_NumJoyAxes      = min(MAXJOYAXES, joystick.numAxes);
-    CONTROL_NumJoyButtons   = min(MAXJOYBUTTONS, joystick.numButtons + 4 * (joystick.numHats > 0));
+    CONTROL_NumJoyAxes      = min((int32_t)MAXJOYAXES, joystick.numAxes);
+    CONTROL_NumJoyButtons   = min((int32_t)MAXJOYBUTTONS, joystick.numButtons + 4 * (joystick.numHats > 0));
     CONTROL_JoystickEnabled = CONTROL_JoyPresent = !!((inputdevices & 4) >> 2);
 
 #ifdef GEKKO
